Use Icey Maze generation, fix bug: invalid bearing.
The "Invalid bearing" error was happening whenever the first tile in the
path between the spawn and goal endpoints was diagonal to the spawn tile.
Luckily, the Dijkstra solver that comes with ROT.js has an option for
defining map topology.

By only considering tiles in each of the four cardinal directions
(instead of all 8 adjacent tiles, which is the default), we can ensure
that all tiles are either to the East, West, North or South of the
previous tile.

Signed-off-by: suicidepills <atom.mac@gmail.com>